chore: update device info api file
